requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
spotipy==2.25.0
python-dotenv==1.0.1
langchain==0.3.13
langchain-core==0.3.26
langchain-openai==0.2.14
langgraph==0.2.60
